rename the create route to slack because that makes more sense
