+++ src/routes.js
@@ -1,6 +1,6 @@
 module.exports = {
   ping: require('./ping'),
   subscriptions: {
-    create: require('./resources/subscriptions/routes/create')
+    slack: require('./resources/subscriptions/routes/slack')
   }
 };
